+++ library-notifications/src/main/java/com/deltadna/android/sdk/notifications/Utils.java
@@ -59,7 +59,6 @@ class Utils {
     static Intent wrapWithReceiver(Context context, Intent intent) {
         synchronized (DDNANotifications.class) {
             if (DDNANotifications.receiver != null) {
-                //noinspection ConstantConditions
                 intent.setClass(context, DDNANotifications.receiver);
             }
         }