+++ library/src/test/java/com/deltadna/android/sdk/DDNAImplTest.kt
@@ -45,6 +45,7 @@ import java.util.concurrent.TimeUnit
 import com.deltadna.android.sdk.EventActionHandler.GameParametersHandler as GPH
 
 @RunWith(RobolectricTestRunner::class)
+@Ignore("Mock server is using outdated payloads and needs to be revisited")
 class DDNAImplTest {
 
     private lateinit var server: MockWebServer